moved all tests for reviews and questions to component specific folders. Adjusted reference paths so tests can run. all tests passing
